tests: move diffusion tests into tests/diffusion/, 14 files into 6
The branch was adding 14 files flat at the top of tests/, where main has 77
entries and two subdirectories (plugin/, entrypoints/) that this work ignored.
Top-level tests/ now gains one directory instead of fourteen files.

Grouped by subject rather than by source module: framework, serving,
attention, then H3's geometry, networks and pipelines. Merged sources keep
their own docstrings under a heavy rule, so the seams stay readable.

Pure reorganisation -- all 224 test functions survive with identical names
(verified by diffing the collected sets), 235 tests pass. Four collisions
resolved by renaming, none of which changes a test: FRAME_ROWS split into
FL2VA_/REF2VA_ (they differed, 1008 vs 24), build() into
build_t2va_pipeline()/build_ref2va_sequence(), and packed_sequence's OBS_SEQ /
OBS_USED dropped as duplicates of geometry's identical values.

The recipe's test count was stale at 247; it is 235.
